FUSETOOLS2-303 - Move to another subfolder to have completion java played before
this is a workaround to have java completion tests working on Travis. It
sounds that if a Java file was opened before, the completion then is not
working, but only on CI...
